insights-1308 continued make the collection work again move spec classes to their own file
